Wait for a yellow status when setting up indexes.
Before, I would get random errors about shards not being available. Erik
Rose told me that this is because I made an index, but then didn't wait
for it to propogate to the shards. He gave me a line of code to wait for
the sync.

I noticed I had to add it in several places, always right after a call
to `setup_indexes()`. So I made that function do that. I also removed
the refresh calls right after any `setup_index()` calls, because refresh
is called in that function.